Add files via upload
It is a online book store application. In this module we want to add or purchase or available of books those details are included. And also it will connected to the database itself. So whenever we want to add the book details in the handler methods that details will be  stored in the database as well as the deleting operations too.